+++ R/tk_augment_timeseries.R
@@ -38,15 +38,9 @@ tk_augment_timeseries_signature.data.frame <- function(data) {
     date_var <- tk_get_timeseries_variables(data)[[1]]
 
     # Arrange by date_var
-    date_var_has_space <- stringr::str_detect("date column", " ")
-    if (date_var_has_space) {
-        data <- data %>%
-            dplyr::arrange_(paste0("`", date_var, "`"))
-    } else {
-        data <- data %>%
-            dplyr::arrange_(date_var)
-    }
+    data <- data %>% dplyr::arrange(!! sym(date_var))
 
+    # Bind Time Series Signature
     ret_1 <- data
 
     ret_2 <- data %>%