+++ tests/testthat/test_tk_xts.R
@@ -1,36 +1,36 @@
 context("Test tk_xts")
 
-AAPL_tbl <- tidyquant::tq_get("AAPL", from = "2015-01-01", to = "2016-12-31")
-AAPL_zoo <- tk_zoo(AAPL_tbl, silent = TRUE)
+FB_tbl <- FANG %>% filter(symbol == "FB")
+FB_zoo <- tk_zoo(FB_tbl, silent = TRUE)
 
 # FUNCTION: tk_xts -----
 
 # tbl to xts -----
 test_that("tbl to xts test returns xts with correct rows and columns.", {
     # Use date column to specify order
-    test_xts_1 <- tk_xts(AAPL_tbl, select = -date, date_var = date)
-    expect_equal(nrow(test_xts_1), 504)
+    test_xts_1 <- tk_xts(FB_tbl, select = -c(date, symbol), date_var = date)
+    expect_equal(nrow(test_xts_1), 1008)
     expect_equal(ncol(test_xts_1), 6)
 
     # Use order.by to specify order
-    test_xts_2 <- tk_xts(AAPL_tbl[,-1], order.by = AAPL_tbl$date)
-    expect_equal(nrow(test_xts_2), 504)
+    test_xts_2 <- tk_xts(FB_tbl[,-2], order.by = FB_tbl$date, silent = TRUE)
+    expect_equal(nrow(test_xts_2), 1008)
     expect_equal(ncol(test_xts_2), 6)
 
     # Auto-index date
-    expect_message(tk_xts(AAPL_tbl, select = -date))  # using `date` column as date
+    expect_message(tk_xts(FB_tbl, select = -c(symbol, date)))  # using `date` column as date
 
     # Auto-drop columns
-    expect_warning(tk_xts(AAPL_tbl, date_var = date)) # dropping date column
+    expect_warning(tk_xts(FB_tbl, date_var = date)) # dropping date column
 
     # Auto-index and auto-drop columns
-    expect_message(expect_warning(tk_xts(AAPL_tbl))) # using `date` column as date, dropping date column
+    expect_message(expect_warning(tk_xts(FB_tbl))) # using `date` column as date, dropping date column
 
     # NSE
     select     <- "adjusted"
     date_var   <- "date"
-    test_xts_3 <- tk_xts_(AAPL_tbl, select = select, date_var = date_var)
-    expect_equal(nrow(test_xts_3), 504)
+    test_xts_3 <- tk_xts_(FB_tbl, select = select, date_var = date_var)
+    expect_equal(nrow(test_xts_3), 1008)
     expect_equal(ncol(test_xts_3), 1)
     expect_equal(colnames(test_xts_3), select)
 
@@ -51,23 +51,23 @@ test_that("tbl to xts test returns xts with correct rows and columns.", {
 # Default is xts::xts() for other objects; only test zoo
 test_that("zoo to xts test returns xts with correct rows and columns.", {
     # Use date column to specify order
-    test_xts_4 <- tk_xts(AAPL_zoo)
-    expect_equal(nrow(test_xts_4), 504)
+    test_xts_4 <- tk_xts(FB_zoo)
+    expect_equal(nrow(test_xts_4), 1008)
     expect_equal(ncol(test_xts_4), 6)
 
     # Use order.by to specify order
-    test_xts_5 <- tk_xts(AAPL_zoo, order.by = zoo::index(AAPL_zoo))
-    expect_equal(nrow(test_xts_5), 504)
+    test_xts_5 <- tk_xts(FB_zoo, order.by = zoo::index(FB_zoo))
+    expect_equal(nrow(test_xts_5), 1008)
     expect_equal(ncol(test_xts_5), 6)
 
     # Warning if using select field
-    expect_warning(tk_xts(AAPL_zoo, select = -date))  # only for use with data.frames
+    expect_warning(tk_xts(FB_zoo, select = -date))  # only for use with data.frames
 
     # Warning if using date_var field
-    expect_warning(tk_xts(AAPL_zoo, date_var = date)) # only for use with data.frames
+    expect_warning(tk_xts(FB_zoo, date_var = date)) # only for use with data.frames
 
     # Two warnings: select and date_var args only applicable to data.frames
-    expect_warning(tk_xts(AAPL_zoo, select = -date, date_var = date))
+    expect_warning(tk_xts(FB_zoo, select = -date, date_var = date))
 
 })
 
@@ -75,9 +75,8 @@ test_that("zoo to xts test returns xts with correct rows and columns.", {
 # Default is xts::xts() for other objects; only test zoo
 test_that("ts to xts test returns xts with correct rows and columns.", {
     # Use date column to specify order
-    test_xts_6 <- tk_ts(AAPL_tbl, select = -date) %>%
-        tk_xts()
-    expect_equal(nrow(test_xts_6), 504)
+    test_xts_6 <- tk_ts(FB_tbl, silent = TRUE) %>% tk_xts()
+    expect_equal(nrow(test_xts_6), 1008)
     expect_equal(ncol(test_xts_6), 6)
 
 })